3763: Group results by data source
 - Refactored to move the APIs added earlier from SleuthkitCase to Blackboard.
